requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
firebase-admin>=6.0.0
tensorflow>=2.20.0
Pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
gunicorn>=21.0.0